requirements.txt
streamlit>=1.44.1
requests>=2.32.3
google-auth>=2.39.0
google-genai>=1.12.1
pydantic>=2.11.3